pyproject.toml
[tool.poetry]
name = "sdss-lvmpwi"
version = "0.0.1"
description = "Lvm Pwi4 Actor"
authors = ["Florian Briegel <briegel@mpia.de>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/lvmpwi"
repository = "https://github.com/sdss/lvmpwi"
documentation = "https://sdss-lvmpwi.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lvmpwi", from = "python" }
]
include = ["setup.py", "python/lvmpwi/etc/*", "config/*", "container/*", "tests/*"]

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[tool.poetry.scripts]
lvmpwi = "lvmpwi.__main__:lvmpwi"
container_build ="container:build"
container_start ="container:start"
container_stop ="container:stop"
container_isRunning ="container:isRunning"

[tool.poetry.dependencies]
sdss-tree = ">=2.15.2"
sdss-access = ">=0.2.3"
sdsstools = ">=0.4.0"
daemonocle = "^1.1.1"
click-default-group = "^1.2.2"
dlipower = "^1.0.176"
sdss-clu = "^1.0.3"
sphinx-bootstrap-theme = "^0.7.1"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = "^3.3.1"
pylint = "^2.8.2"
Sphinx = {version="^4.0.2", optional=true}
sphinx_bootstrap_theme = { version=">=0.4.12", optional=true}

[tool.poetry.extras]
docs = ["Sphinx", "sphinx_bootstrap_theme"]

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "lvmpwi"
known_sdss = ["sdsstools"]
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
#addopts = "--cov lvmpwi --cov-report xml --cov-report html --cov-report term"
testpaths = [
    "tests",
]


[tool.coverage.run]
branch = true
include = ["python/lvmpwi/*"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=49.6.0"]
build-backend = "poetry.core.masonry.api"
